Better roundtrip test for single-component cps
