+++ test/integration/cps/test_cps.py
@@ -254,3 +254,32 @@ def test_extended_cpp_info():
     assert pkg_comp["location"] == "my_custom_location"
     assert pkg_comp["link_languages"] == ["cpp"]
     assert pkg_comp["definitions"] == {'cpp': {'MY_DEFINE': None, 'MY_OTHER_DEFINE': '1'}}
+
+
+def test_cps_component_single():
+    c = TestClient()
+    conanfile = textwrap.dedent("""\
+        from conan import ConanFile
+        from conan.tools.files import save
+        import os
+
+        class mypkgRecipe(ConanFile):
+            name = "mypkg"
+            version = "0.1"
+
+            def package(self):
+                save(self, os.path.join(self.package_folder, "lib", "libcore.a"), "")
+
+            def package_info(self):
+                self.cpp_info.components["core"].libs = ["core"]
+                from conan.cps import CPS
+                cps = CPS.from_conan(self)
+                self.cpp_info = cps.to_conan()
+        """)
+    c.save({"conanfile.py": conanfile})
+    c.run("create")
+    c.run(f"install --requires=mypkg/0.1 -g CMakeConfigDeps")
+    mypkg_targets = c.load("mypkg-Targets-release.cmake")
+    assert "add_library(mypkg::core" in mypkg_targets
+    # But we don't create what would otherwise be the default component
+    assert "add_library(mypkg::mypkg" not in mypkg_targets
